+++ packages/core/src/domain/session/sessionStore.spec.ts
@@ -69,6 +69,7 @@ function getStoreExpiration() {
 
 function resetSessionInStore() {
   sessionStoreStrategy.expireSession()
+  sessionStoreStrategy.expireSession.calls.reset()
 }
 
 function setSessionInStore(sessionState: SessionState) {
@@ -210,6 +211,7 @@ describe('session store', () => {
         computeTrackingType,
         sessionStoreStrategy
       )
+      sessionStoreStrategy.persistSession.calls.reset()
       sessionStoreManager.expireObservable.subscribe(expireSpy)
       sessionStoreManager.renewObservable.subscribe(renewSpy)
     }
@@ -474,17 +476,6 @@ describe('session store', () => {
         expect(sessionStoreStrategy.persistSession).toHaveBeenCalled()
       })
 
-      it('when session not in cache and session in store, should do nothing', () => {
-        setupSessionStore()
-        setSessionInStore(createSessionState(FakeTrackingType.TRACKED, FIRST_ID))
-
-        clock.tick(STORAGE_POLL_DELAY)
-
-        expect(sessionStoreManager.getSession().id).toBeUndefined()
-        expect(expireSpy).not.toHaveBeenCalled()
-        expect(sessionStoreStrategy.persistSession).not.toHaveBeenCalled()
-      })
-
       it('when session in cache and session not in store, should expire session', () => {
         setupSessionStore(createSessionState(FakeTrackingType.TRACKED, FIRST_ID))
         resetSessionInStore()
@@ -495,6 +486,17 @@ describe('session store', () => {
         expectSessionToBeExpiredInStore()
         expect(expireSpy).toHaveBeenCalled()
         expect(sessionStoreStrategy.persistSession).toHaveBeenCalled()
+      })
+
+      it('when session not in cache and session in store, should do nothing', () => {
+        setupSessionStore()
+        setSessionInStore(createSessionState(FakeTrackingType.TRACKED, FIRST_ID))
+
+        clock.tick(STORAGE_POLL_DELAY)
+
+        expect(sessionStoreManager.getSession().id).toBeUndefined()
+        expect(expireSpy).not.toHaveBeenCalled()
+        expect(sessionStoreStrategy.persistSession).not.toHaveBeenCalled()
       })
 
       it('when session in cache is same session than in store, should synchronize session', () => {
@@ -520,6 +522,26 @@ describe('session store', () => {
         expect(sessionStoreManager.getSession().id).toBeUndefined()
         expect(expireSpy).toHaveBeenCalled()
         expect(sessionStoreStrategy.persistSession).not.toHaveBeenCalled()
+      })
+
+      it('when session in store is expired first and then get updated by another tab, should expire session in cache and not touch the store', () => {
+        setupSessionStore(createSessionState(FakeTrackingType.TRACKED, FIRST_ID))
+        resetSessionInStore()
+
+        // Simulate a new session being written to the store by another tab during the watch.
+        // Watch is reading the cookie twice so we need to plan the write of the cookie at the right index
+        sessionStoreStrategy.planRetrieveSession(1, createSessionState(FakeTrackingType.TRACKED, SECOND_ID))
+
+        clock.tick(STORAGE_POLL_DELAY)
+
+        // expires session in cache
+        expect(sessionStoreManager.getSession().id).toBeUndefined()
+        expect(expireSpy).toHaveBeenCalled()
+
+        // Does not touch the store
+        // The two calls to persist session are for the lock management, these can be ignored
+        expect(sessionStoreStrategy.persistSession).toHaveBeenCalledTimes(2)
+        expect(sessionStoreStrategy.expireSession).not.toHaveBeenCalled()
       })
 
       it('when session type in cache is different than session type in store, should expire session and not touch the store', () => {