+++ packages/rum/src/domain/segmentCollection/index.ts
@@ -1,3 +1,4 @@
 export { startSegmentCollection, setSegmentBytesLimit } from './segmentCollection'
 export { SEGMENT_BYTES_LIMIT } from './segmentCollection'
-export type { BrowserSegmentMetadataAndSegmentSizes } from './buildReplayPayload'
+export type { BrowserSegmentMetadataAndSegmentSizes, ReplayPayload } from './buildReplayPayload'
+export { startSegmentTelemetry } from './startSegmentTelemetry'