+++ packages/rum/src/domain/segmentCollection/segment.spec.ts
@@ -1,4 +1,4 @@
-import type { DeflateEncoder, TimeStamp } from '@datadog/browser-core'
+import type { DeflateEncoder, TimeStamp, Uint8ArrayBuffer } from '@datadog/browser-core'
 import { noop, setDebugMode, DeflateEncoderStreamId } from '@datadog/browser-core'
 import type { RumConfiguration } from '@datadog/browser-rum-core'
 import { registerCleanupTask } from '@datadog/browser-core/test'
@@ -7,12 +7,14 @@ import type { CreationReason, BrowserRecord, SegmentContext, BrowserSegment, Bro
 import { RecordType } from '../../types'
 import { getReplayStats, resetReplayStats } from '../replayStats'
 import { createDeflateEncoder } from '../deflate'
+import type { SerializationStats } from '../record'
 import type { AddRecordCallback, FlushCallback, Segment } from './segment'
 import { createSegment } from './segment'
 
 const CONTEXT: SegmentContext = { application: { id: 'a' }, view: { id: 'b' }, session: { id: 'c' } }
 const RECORD_TIMESTAMP = 10 as TimeStamp
 const RECORD: BrowserRecord = { type: RecordType.ViewEnd, timestamp: RECORD_TIMESTAMP }
+const RECORD_STATS = { cssText: { count: 0, max: 0, sum: 0 } }
 const FULL_SNAPSHOT_RECORD: BrowserRecord = {
   type: RecordType.FullSnapshot,
   timestamp: RECORD_TIMESTAMP,
@@ -43,7 +45,7 @@ describe('Segment', () => {
     const addRecordCallbackSpy = jasmine.createSpy<AddRecordCallback>()
     const flushCallbackSpy = jasmine.createSpy<FlushCallback>()
     const segment = createTestSegment()
-    segment.addRecord(RECORD, addRecordCallbackSpy)
+    segment.addRecord(RECORD, RECORD_STATS, addRecordCallbackSpy)
 
     worker.processAllMessages()
     expect(addRecordCallbackSpy).toHaveBeenCalledTimes(1)
@@ -54,7 +56,7 @@ describe('Segment', () => {
     expect(addRecordCallbackSpy).toHaveBeenCalledTimes(1)
     expect(flushCallbackSpy).toHaveBeenCalledTimes(1)
 
-    expect(parseSegment(flushCallbackSpy.calls.mostRecent().args[1].output)).toEqual({
+    expect(parseSegment(flushCallbackSpy.calls.mostRecent().args[2].output)).toEqual({
       source: 'browser' as const,
       creation_reason: 'init' as const,
       end: 10,
@@ -75,7 +77,7 @@ describe('Segment', () => {
   it('compressed bytes count is updated when a record is added', () => {
     const addRecordCallbackSpy = jasmine.createSpy<AddRecordCallback>()
     const segment = createTestSegment()
-    segment.addRecord(RECORD, addRecordCallbackSpy)
+    segment.addRecord(RECORD, RECORD_STATS, addRecordCallbackSpy)
     worker.processAllMessages()
     expect(addRecordCallbackSpy).toHaveBeenCalledOnceWith(
       ENCODED_SEGMENT_HEADER_BYTES_COUNT + ENCODED_RECORD_BYTES_COUNT
@@ -85,7 +87,7 @@ describe('Segment', () => {
   it('calls the flush callback with metadata and encoder output as argument', () => {
     const flushCallbackSpy = jasmine.createSpy<FlushCallback>()
     const segment = createTestSegment()
-    segment.addRecord(RECORD, noop)
+    segment.addRecord(RECORD, RECORD_STATS, noop)
     segment.flush(flushCallbackSpy)
     worker.processAllMessages()
     expect(flushCallbackSpy).toHaveBeenCalledOnceWith(
@@ -99,8 +101,9 @@ describe('Segment', () => {
         records_count: 1,
         ...CONTEXT,
       },
+      RECORD_STATS,
       {
-        output: jasmine.any(Uint8Array) as unknown as Uint8Array,
+        output: jasmine.any(Uint8Array) as unknown as Uint8ArrayBuffer,
         outputBytesCount:
           ENCODED_SEGMENT_HEADER_BYTES_COUNT +
           ENCODED_RECORD_BYTES_COUNT +
@@ -116,16 +119,20 @@ describe('Segment', () => {
     const flushCallbackSpy = jasmine.createSpy<FlushCallback>()
 
     const segment1 = createTestSegment({ creationReason: 'init' })
-    segment1.addRecord(RECORD, noop)
+    const stats1 = { cssText: { count: 1, max: 100, sum: 150 } }
+    segment1.addRecord(RECORD, stats1, noop)
     segment1.flush(flushCallbackSpy)
 
     const segment2 = createTestSegment({ creationReason: 'segment_duration_limit' })
-    segment2.addRecord(FULL_SNAPSHOT_RECORD, noop)
+    const stats2 = { cssText: { count: 2, max: 200, sum: 275 } }
+    segment2.addRecord(FULL_SNAPSHOT_RECORD, stats2, noop)
     segment2.flush(flushCallbackSpy)
 
     worker.processAllMessages()
-    expect(parseSegment(flushCallbackSpy.calls.argsFor(0)[1].output).records.length).toBe(1)
-    expect(parseSegment(flushCallbackSpy.calls.argsFor(1)[1].output).records.length).toBe(1)
+    expect(flushCallbackSpy.calls.argsFor(0)[1]).toEqual(stats1)
+    expect(parseSegment(flushCallbackSpy.calls.argsFor(0)[2].output).records.length).toBe(1)
+    expect(flushCallbackSpy.calls.argsFor(1)[1]).toEqual(stats2)
+    expect(parseSegment(flushCallbackSpy.calls.argsFor(1)[2].output).records.length).toBe(1)
   })
 
   it('throws when trying to flush an empty segment', () => {
@@ -138,8 +145,8 @@ describe('Segment', () => {
       let segment: Segment
       beforeEach(() => {
         segment = createTestSegment()
-        segment.addRecord({ type: RecordType.ViewEnd, timestamp: 10 as TimeStamp }, noop)
-        segment.addRecord({ type: RecordType.ViewEnd, timestamp: 15 as TimeStamp }, noop)
+        segment.addRecord({ type: RecordType.ViewEnd, timestamp: 10 as TimeStamp }, RECORD_STATS, noop)
+        segment.addRecord({ type: RecordType.ViewEnd, timestamp: 15 as TimeStamp }, RECORD_STATS, noop)
       })
 
       it('does increment records_count', () => {
@@ -151,7 +158,7 @@ describe('Segment', () => {
       })
 
       it('does change the start timestamp when receiving an earlier record', () => {
-        segment.addRecord({ type: RecordType.ViewEnd, timestamp: 5 as TimeStamp }, noop)
+        segment.addRecord({ type: RecordType.ViewEnd, timestamp: 5 as TimeStamp }, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment).start).toBe(5)
       })
 
@@ -160,7 +167,7 @@ describe('Segment', () => {
       })
 
       it('does not change the end timestamp when receiving an earlier record', () => {
-        segment.addRecord({ type: RecordType.ViewEnd, timestamp: 5 as TimeStamp }, noop)
+        segment.addRecord({ type: RecordType.ViewEnd, timestamp: 5 as TimeStamp }, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment).end).toBe(15)
       })
     })
@@ -168,20 +175,20 @@ describe('Segment', () => {
     describe('has_full_snapshot', () => {
       it('sets has_full_snapshot to false if a segment has a no FullSnapshot', () => {
         const segment = createTestSegment()
-        segment.addRecord(RECORD, noop)
+        segment.addRecord(RECORD, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment).has_full_snapshot).toEqual(false)
       })
 
       it('sets has_full_snapshot to true if a segment has a FullSnapshot', () => {
         const segment = createTestSegment()
-        segment.addRecord(FULL_SNAPSHOT_RECORD, noop)
+        segment.addRecord(FULL_SNAPSHOT_RECORD, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment).has_full_snapshot).toEqual(true)
       })
 
       it("doesn't overrides has_full_snapshot to false once it has been set to true", () => {
         const segment = createTestSegment()
-        segment.addRecord(FULL_SNAPSHOT_RECORD, noop)
-        segment.addRecord(RECORD, noop)
+        segment.addRecord(FULL_SNAPSHOT_RECORD, RECORD_STATS, noop)
+        segment.addRecord(RECORD, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment).has_full_snapshot).toEqual(true)
       })
     })
@@ -189,25 +196,25 @@ describe('Segment', () => {
     describe('index_in_view', () => {
       it('increments index_in_view every time a segment is created for the same view', () => {
         const segment1 = createTestSegment()
-        segment1.addRecord(RECORD, noop)
+        segment1.addRecord(RECORD, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment1).index_in_view).toBe(0)
 
         const segment2 = createTestSegment()
-        segment2.addRecord(RECORD, noop)
+        segment2.addRecord(RECORD, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment2).index_in_view).toBe(1)
 
         const segment3 = createTestSegment()
-        segment3.addRecord(RECORD, noop)
+        segment3.addRecord(RECORD, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment3).index_in_view).toBe(2)
       })
 
       it('resets segments_count when creating a segment for a new view', () => {
         const segment1 = createTestSegment()
-        segment1.addRecord(RECORD, noop)
+        segment1.addRecord(RECORD, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment1).index_in_view).toBe(0)
 
         const segment2 = createTestSegment({ context: { ...CONTEXT, view: { id: 'view-2' } } })
-        segment2.addRecord(RECORD, noop)
+        segment2.addRecord(RECORD, RECORD_STATS, noop)
         expect(flushAndGetMetadata(segment2).index_in_view).toBe(0)
       })
     })
@@ -219,6 +226,51 @@ describe('Segment', () => {
       })
       worker.processAllMessages()
       return metadata!
+    }
+  })
+
+  describe('serialization stats', () => {
+    describe('when adding a record', () => {
+      let segment: Segment
+      beforeEach(() => {
+        segment = createTestSegment()
+        segment.addRecord(
+          { type: RecordType.ViewEnd, timestamp: 10 as TimeStamp },
+          {
+            cssText: { count: 1, max: 50, sum: 50 },
+          },
+          noop
+        )
+        segment.addRecord(
+          { type: RecordType.ViewEnd, timestamp: 15 as TimeStamp },
+          {
+            cssText: { count: 2, max: 150, sum: 250 },
+          },
+          noop
+        )
+      })
+
+      it('aggregates stats', () => {
+        expect(flushAndGetSerializationStats(segment)).toEqual({
+          cssText: { count: 3, max: 150, sum: 300 },
+        })
+      })
+
+      it('does not change aggregated stats for records with no serialization stats', () => {
+        segment.addRecord({ type: RecordType.ViewEnd, timestamp: 5 as TimeStamp }, undefined, noop)
+        expect(flushAndGetSerializationStats(segment)).toEqual({
+          cssText: { count: 3, max: 150, sum: 300 },
+        })
+      })
+    })
+
+    function flushAndGetSerializationStats(segment: Segment) {
+      let stats: SerializationStats
+      segment.flush((_metadata, _stats) => {
+        stats = _stats
+      })
+      worker.processAllMessages()
+      return stats!
     }
   })
 
@@ -241,7 +293,7 @@ describe('Segment', () => {
 
     it('when flushing a segment', () => {
       const segment = createTestSegment()
-      segment.addRecord(RECORD, noop)
+      segment.addRecord(RECORD, RECORD_STATS, noop)
       segment.flush(noop)
       worker.processAllMessages()
       expect(getReplayStats('b')).toEqual(