+++ test/e2e/lib/framework/pageSetups.ts
@@ -1,6 +1,6 @@
 import { generateUUID, INTAKE_URL_PARAMETERS } from '@datadog/browser-core'
 import type { LogsInitConfiguration } from '@datadog/browser-logs'
-import type { RumInitConfiguration } from '@datadog/browser-rum-core'
+import type { RumInitConfiguration, RemoteConfiguration } from '@datadog/browser-rum-core'
 import type test from '@playwright/test'
 import type { Servers } from './httpServers'
 
@@ -10,6 +10,7 @@ export interface SetupOptions {
   logs?: LogsInitConfiguration
   logsInit: (initConfiguration: LogsInitConfiguration) => void
   rumInit: (initConfiguration: RumInitConfiguration) => void
+  remoteConfiguration?: RemoteConfiguration
   eventBridge: boolean
   head?: string
   body?: string
@@ -158,7 +159,7 @@ export function npmSetup(options: SetupOptions, servers: Servers) {
   })
 }
 
-export function reactSetup(options: SetupOptions, servers: Servers) {
+export function reactSetup(options: SetupOptions, servers: Servers, appName: string) {
   let header = options.head || ''
   let body = options.body || ''
 
@@ -175,7 +176,7 @@ export function reactSetup(options: SetupOptions, servers: Servers) {
     `
   }
 
-  body += html` <script type="text/javascript" src="./react-app.js"></script> `
+  body += html` <script type="text/javascript" src="./${appName}.js"></script> `
 
   return basePage({
     header,
@@ -243,6 +244,7 @@ function formatConfiguration(initConfiguration: LogsInitConfiguration | RumInitC
     {
       ...initConfiguration,
       proxy: servers.intake.url,
+      remoteConfigurationProxy: `${servers.base.url}/config`,
     },
     (_key, value) => {
       if (typeof value === 'function') {